fixing Task2 compilation error on master

+++ Task2/Services/EstateTransactionServices.fs
@@ -15,7 +15,7 @@ module EstateTransactionServices =
         0.0
 
     let GetAveragePricePerSquareFeetByCity (transactions: seq<EstateTransaction>)  =
-        Map.empty<string, float>
+        Map.empty<string, decimal>
